Consistently rename requiredddata to required_data
