+++ nomenclature/processor/__init__.py
@@ -2,4 +2,4 @@ from nomenclature.processor.region import (  # noqa
     RegionAggregationMapping,
     RegionProcessor,
 )
-from nomenclature.processor.requireddata import RequiredDataValidator  # noqa
+from nomenclature.processor.required_data import RequiredDataValidator  # noqa
